requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
